requirements.txt
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==1.11.29
django-contrib-comments==1.9.2
django-cors-headers==3.2.1
django-extensions==2.2.8
djangorestframework==3.11.0
entrypoints==0.3
et-xmlfile==1.0.1
fcm-django==1.0.5
filebrowser-safe==0.5.0
firebase-admin==5.1.0
future==0.18.2
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grappelli-safe==0.5.2
grpcio==1.41.1
grpcio-status==1.41.1
httplib2==0.20.2
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
lxml==4.6.4
MarkupSafe==1.1.1
Mezzanine==4.3.1
mistune==0.8.4
msgpack==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.21.4
oauthlib==3.1.0
openpyxl==3.0.4
packaging==21.2
pandas==1.3.4
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plivo==4.3.8
prometheus-client==0.7.1
prompt-toolkit==3.0.4
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pyzmq==22.3.0
qtconsole==4.7.1
QtPy==1.9.0
reportlab==3.6.2
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.0.0
uritemplate==4.1.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
XlsxWriter==1.2.9
zipp==3.1.0
